CRM/backend/requirements-test.txt
# Testing dependencies for Accountant CRM

# Core testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.2.0

# Additional testing utilities
factory-boy>=3.3.0
faker>=19.0.0
coverage>=7.3.0

# HTTP mocking
responses>=0.23.0
requests-mock>=1.11.0

# Async testing (if needed)
pytest-asyncio>=0.21.0

# Database testing
pytest-postgresql>=5.0.0

# Time mocking
freezegun>=1.2.0

# Install with:
# pip install -r requirements-test.txt
